Data Loading script by Erik
Roberto has already split the testing  data for us, so we only need to process the validation data and training data.   Check the source where we download the raw data for explanation!

Hence, I recreated the folder for organizing all the raw data. Also, based on this, I rewrote the python script, and stored the process as followed in the folder.